Cap cyborg-tempest-plugin version for stable/ussuri|victoria
We are capping stable/ussuri and victoria testing with Tempest
26.1.0 (choosing 26.1.0 because of bug#1955418)

- https://review.opendev.org/c/openstack/devstack/+/838051
- https://review.opendev.org/c/openstack/devstack/+/843295

and cyborg-tempest jobs in stable/ussuri|victoria is
failing due to that

- https://zuul.opendev.org/t/openstack/build/b5c679bb6ddf4b7cae630b8cc2b90766
To fix that we need to use compatible cyborg-tempest-plugin version for
Tempest 26.1.0, which is cyborg-tempest-plugin 1.4.0.

As stable/victoria is in EM, droping the victoria job from master gate.

Depends-On: https://review.opendev.org/c/openstack/devstack/+/838051

Change-Id: I8c599b7f8bb4705c12f67729c430aff2c047bc54